+++ src/datasets/german_dataset.py
@@ -77,7 +77,8 @@ class GermanDataset(Dataset):
             'foreign_worker': CAT,
             'class': CAT
         }
-        
+
+        data = data.astype(float)
         data.reset_index(drop=True, inplace=True)
 
         super().__init__(data, sensitive_attrs, target_attr, privileged_class, feature_types,
